refactor(#319): unify CLI grammar; group maintenance verbs under a sub-app
Hard break (no deprecated aliases), per the operator's call on #319:

- New 'nhf-targets maintenance' sub-app hosting check-config (was
  upgrade-config), check-manifest (was upgrade-manifest),
  rebuild-manifest, and rechunk. The check-* rename makes the
  report-only vs mutating split a grammar property.
- -d short alias for --project-dir on every fetch and agg command via a
  shared _PROJECT_DIR_PARAM in cli/_params.py (release's copy folded in).
- release publish --scope sources (was the singular 'source'); the
  source-key flag is now --source everywhere (was --source-key).
  Behavior unchanged: a confirmed sources publish still requires
  --source <key>.
- run --target accepts the pixi-task nicknames rch/som/sca/swe.
- 'aggregate' accepted as an alias of the agg sub-app; app tagline now
  says nhf-targets (matching the binary).

pixi tasks upgrade-config/upgrade-manifest renamed to
check-config/check-manifest; rebuild-manifest and rechunk task names
unchanged but now dispatch through the maintenance sub-app. SLURM
rechunk scripts, CLAUDE.md, README, CONTRIBUTING, docs/maintenance.md,
api/architecture/data-release pages updated in the same PR (historical
specs/plans/reviews left as point-in-time records).

Closes #319
